Rename _ImageChunk to UfuncArrayProcessor and split into ufunc module
Also expanded docstring and reorganized methods to approximate
execution order, plus some minor unrelated refactoring.
